Modified the database structure
Now there is another collection - a country collection. Each country has
its ISO2 Code as an ID, Name and a time zone.

Due to that change the structure of postring/getting the bank data also
had to change - in particular the changes required were to insert a
country while inserting a new bank, and getting the country name for a
bank. Still this small modification does not call for a relational
database, since the countries document is relatively small.